ai-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
tensorflow==2.15.0
keras==2.15.0
pillow==10.1.0
opencv-python==4.8.1.78
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0